Update protobuf from 3.1.0 to 3.2.1. CONTRIB-139
Divergence from upstream:

- Keep using old style maps.

See https://st.yandex-team.ru/CONTRIB-139#1514371763000

- Do not mark RepeatedField and RepeatedPtrField as final to support IsIn util helper.

Fixes https://ci.yandex-team.ru/check/eyjx

- Fix python crash on access to private getters of google.protobuf.pyext._message.Message.

Fixes https://ci.yandex-team.ru/check/eynv
Upstream issue: https://github.com/google/protobuf/issues/2974 (introduced in 3.2.0, not fixed yet).
Tested by the new contrib/libs/protobuf/python/test

ref:58f3a43b33460b885f6760267ba62e7c5f9f304d

+++ contrib/libs/protobuf/compiler/java/java_enum_field.cc
@@ -67,7 +67,8 @@ void SetEnumVariables(const FieldDescriptor* descriptor,
   (*variables)["default"] = ImmutableDefaultValue(descriptor, name_resolver);
   (*variables)["default_number"] = SimpleItoa(
       descriptor->default_value_enum()->number());
-  (*variables)["tag"] = SimpleItoa(internal::WireFormat::MakeTag(descriptor));
+  (*variables)["tag"] =
+      SimpleItoa(static_cast<int32>(internal::WireFormat::MakeTag(descriptor)));
   (*variables)["tag_size"] = SimpleItoa(
       internal::WireFormat::TagSize(descriptor->number(), GetType(descriptor)));
   // TODO(birdo): Add @deprecated javadoc when generating javadoc is supported
